MLK-14737: memtool: cleanup: Fix int-to-pointer-cast warnings
This patch fixes the folowing warnings by changing the
type of g_map_vaddr to uintptr_t which is guaranteed to be able to hold
a pointer.

memtool.c: In function 'map_address':
memtool.c:78:10: warning: cast to pointer from integer of different size
[-Wint-to-pointer-cast]
   munmap((void *)g_map_vaddr, MAP_SIZE);
          ^
memtool.c:81:6: warning: cast from pointer to integer of different size
[-Wpointer-to-int-cast]
      (int)mmap(NULL, MAP_SIZE, PROT_READ | PROT_WRITE, MAP_SHARED,
      ^
memtool.c: In function 'readm':
memtool.c:97:11: warning: cast to pointer from integer of different size
[-Wint-to-pointer-cast]
  addr32 = (uint32_t *) (g_map_vaddr + (address & (MAP_SIZE - 1)));
           ^
memtool.c: In function 'writem':
memtool.c:122:11: warning: cast to pointer from integer of different
size [-Wint-to-pointer-cast]
  addr32 = (uint32_t *) (g_map_vaddr + (address & (MAP_SIZE - 1)));
           ^
memtool.c: In function 'main':
memtool.c:680:10: warning: cast to pointer from integer of different
size [-Wint-to-pointer-cast]
   munmap((void *)g_map_vaddr, MAP_SIZE);

+++ test/memtool/memtool.c
@@ -36,7 +36,7 @@ char *g_field;
 char *g_reg_input;
 
 int g_fmem;
-int g_map_vaddr = 0;
+uintptr_t g_map_vaddr = 0;
 int g_map_paddr = 0;
 int g_comp = 0;
 int g_module_match = 0;
@@ -78,7 +78,7 @@ int map_address(int address)
 		munmap((void *)g_map_vaddr, MAP_SIZE);
 
 	g_map_vaddr =
-	    (int)mmap(NULL, MAP_SIZE, PROT_READ | PROT_WRITE, MAP_SHARED,
+	    (uintptr_t)mmap(NULL, MAP_SIZE, PROT_READ | PROT_WRITE, MAP_SHARED,
 		      g_fmem, address & (~(MAP_SIZE - 1)));
 	g_map_paddr = address & (~(MAP_SIZE - 1));
